Finish detail screen

+++ app/src/main/java/com/example/events/data/Event.kt
@@ -5,8 +5,8 @@ data class Event(
     val title: String,
     val description: String = "",
     val image:String? = null,
-    val longitude: Float? = null,
-    val latitude: Float? = null,
+    val longitude: Double? = null,
+    val latitude: Double? = null,
     val price: String = "",
     val date: Long? = null,
     val people: List<CheckIn> = listOf()
